requirements.txt
requests==2.32.5
spacetrack==1.4.0
pyyaml==6.0.2
python-dotenv==1.2.1
sgp4==2.25
numpy==1.26.4
fastapi==0.128.0
uvicorn==0.40.0
reportlab==4.4.9
